Refactored handlers.go into another file tornjak_handlers.go
Refactored handlers.go
