+++ api/agent/handlers.go
@@ -733,286 +733,3 @@ func (s *Server) federationDelete(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 }
-
-// Tornjak Handlers
-func (s *Server) home(w http.ResponseWriter, r *http.Request) {
-	var ret = "Welcome to the Tornjak Backend!"
-
-	cors(w, r)
-	je := json.NewEncoder(w)
-
-	var err = je.Encode(ret)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-	}
-}
-
-func (s *Server) health(w http.ResponseWriter, r *http.Request) {
-	var ret = "Endpoint is healthy."
-
-	cors(w, r)
-	je := json.NewEncoder(w)
-
-	var err = je.Encode(ret)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-	}
-}
-
-func (s *Server) tornjakSelectorsList(w http.ResponseWriter, r *http.Request) {
-	buf := new(strings.Builder)
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-	var input ListSelectorsRequest
-	if n == 0 {
-		input = ListSelectorsRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-	ret, err := s.ListSelectors(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	je := json.NewEncoder(w)
-	err = je.Encode(ret)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-}
-
-func (s *Server) tornjakPluginDefine(w http.ResponseWriter, r *http.Request) {
-	buf := new(strings.Builder)
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-	var input RegisterSelectorRequest
-	if n == 0 {
-		input = RegisterSelectorRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-	err = s.DefineSelectors(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	_, err = w.Write([]byte("SUCCESS"))
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-}
-
-func (s *Server) tornjakAgentsList(w http.ResponseWriter, r *http.Request) {
-	buf := new(strings.Builder)
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-	var input ListAgentMetadataRequest
-	if n == 0 {
-		input = ListAgentMetadataRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-	ret, err := s.ListAgentMetadata(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	je := json.NewEncoder(w)
-	err = je.Encode(ret)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-}
-
-/********* CLUSTER *********/
-
-func (s *Server) clusterList(w http.ResponseWriter, r *http.Request) {
-	var input ListClustersRequest
-	buf := new(strings.Builder)
-
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-
-	if n == 0 {
-		input = ListClustersRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-
-	ret, err := s.ListClusters(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	je := json.NewEncoder(w)
-	err = je.Encode(ret)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-}
-
-func (s *Server) clusterCreate(w http.ResponseWriter, r *http.Request) {
-	buf := new(strings.Builder)
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-	var input RegisterClusterRequest
-	if n == 0 {
-		input = RegisterClusterRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-	err = s.DefineCluster(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	_, err = w.Write([]byte("SUCCESS"))
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-}
-
-func (s *Server) clusterEdit(w http.ResponseWriter, r *http.Request) {
-	buf := new(strings.Builder)
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-	var input EditClusterRequest
-	if n == 0 {
-		input = EditClusterRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-	err = s.EditCluster(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	_, err = w.Write([]byte("SUCCESS"))
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-}
-
-func (s *Server) clusterDelete(w http.ResponseWriter, r *http.Request) {
-	buf := new(strings.Builder)
-	n, err := io.Copy(buf, r.Body)
-	if err != nil {
-		emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	data := buf.String()
-	var input DeleteClusterRequest
-	if n == 0 {
-		input = DeleteClusterRequest{}
-	} else {
-		err := json.Unmarshal([]byte(data), &input)
-		if err != nil {
-			emsg := fmt.Sprintf("Error parsing data: %v", err.Error())
-			retError(w, emsg, http.StatusBadRequest)
-			return
-		}
-	}
-	err = s.DeleteCluster(input)
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-	cors(w, r)
-	_, err = w.Write([]byte("SUCCESS"))
-	if err != nil {
-		emsg := fmt.Sprintf("Error: %v", err.Error())
-		retError(w, emsg, http.StatusBadRequest)
-		return
-	}
-
-}
-
-/********* END CLUSTER *********/
